refactor: various corrections including phpcs
Refs #6.0.x
