+++ src/Flash/AbstractFlash.php
@@ -16,7 +16,7 @@ namespace Phalcon\Flash;
 use Phalcon\Di\InjectionAwareInterface;
 use Phalcon\Di\Traits\InjectionAwareTrait;
 use Phalcon\Flash\Traits\FlashGettersTrait;
-use Phalcon\Html\EscaperInterface;
+use Phalcon\Html\Escaper\EscaperInterface;
 use Phalcon\Session\ManagerInterface as SessionInterface;
 use Phalcon\Traits\Helper\Str\InterpolateTrait;
