+++ src/Html/Escaper/Exception.php
@@ -1,0 +1,20 @@
+<?php
+
+/**
+ * This file is part of the Phalcon Framework.
+ *
+ * For the full copyright and license information, please view the LICENSE.md
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Phalcon\Html\Escaper;
+
+/**
+ * Class Exception
+ */
+class Exception extends \Exception
+{
+
+}
